+++ changelog.md
@@ -325,5 +325,8 @@
     //    jw, imdb: fixed not working when no rating present
     // ver 0.6.12 @ 2023-7-29
     //    wavve: supports /supermultisection
-    // ver 0.6.13 @ 2023-8-6
+    // ver 0.6.16 @ 2023-8-7
     //    kino: if searching failed and imdb rating already present on page, use the present rating
+    //    cache: changed some key names for...
+    //    watcha: improved search (additional searching) when org. title is available
+    //    internal: clean-up